Move sdk/go/common/resource/testing to pkg
