+++ pkg/resource/testing/rapid.go
